chore: improve defaults in export

+++ plantseg/viewer_napari/widgets/io.py
@@ -224,7 +224,7 @@ def widget_export_image(
     export_format: str = "tiff",
     key: str = "raw",
     scale_to_origin: bool = True,
-    data_type: str = "float32",
+    data_type: str = "uint16",
 ) -> None:
     timer = time.time()
     log("export_image_task started", thread="Export stacks", level="info")
